Привел результат к int

+++ src/test/java/ru/itpark/CreditServiceTest.java
@@ -10,13 +10,13 @@ class CreditServiceTest {
     void calculateCredit() {
         {
             CreditService creditService = new CreditService();
-            double v = creditService.calculateCredit(100_000, 12, 12);
+            int v = creditService.calculateCredit(100_000, 12, 12);
             assertEquals(106619, v);
         }
 
         {
             CreditService creditService = new CreditService();
-            double v = creditService.calculateCredit(100_000_000, 7.5, 12);
+            int v = creditService.calculateCredit(100_000_000, 7.5, 12);
             assertEquals(104_108_900, v);
         }
